(gha) Test install in gha on other os's using containers
Github actions only provides a few ubuntu vms. This patch takes a stab
at adding a matrix of base os containers to run on ubuntu 24.04 in an
attempt to simulate install on those variants.

To simplify the workflow, I'm not attempting to install bolt as a runner
for these additional cases. Instead, the facts module is checked out
alongside and I simulate the run environment by supplying the Bolt
PT__installdir var (and other PT_* task vars) to the task script
executed directly.

This does produce duplicates of the defaults for
apt_source/yum_source/collection, which isn't great...