(tasks) Add defaults to install_build_artifact_linux.sh
For the PT_package and PT_artifacts_source variables. These defaults are
provided by Bolt when the task is run by Bolt. However they need to come
form somewhere when the script is executing on its own, as it is in the
GHA pr_testing_install_build_artifact.yaml workflow. So the question
becomes whether to duplicate defaults in the script or the gha workflow,
and I chose the script.