(tasks) Translate Debian codename to version when version is n/a
The /etc/os-release data for Debian pre-release builds only contains the
codename, not the version number. The VERSION_ID strings are all 'n/a'.

This is a problem for installing openvox packages, because their naming
scheme includes the release number.

This patch adds a little static stranslation of codename to version when
we encounter an 'n/a' version.

This isn't the greatest solution, since it will need ongoing upkeep for
future Debian releases, but it is self contained.

I considered accepting a version override as a task parameter, but it's
not simple for the caller to be able to consistently provide this
information either.

It's unknown to me whether the next Ubuntu LTS release will have a similar
problem, but I haven't found a 26.04 codename either.