(tasks) Ensure rpm install is idempotent
By default, rpm will fail if asked to install a package file that has
already been installed. This would cause the install task to fail if
re-run unless the openvox release package happened to have been updated
since the last execution.

Adding --replacepkgs to the rpm call allows it to reinstall the release
package.

Also added an idempotency check to the gha workflow by re-running bolt
or the script to verify that it doesn't fail run a second time.