(maint) Prefer apt-get to apt
For the couple of package manager operations in common.sh, apt had
preference before apt-get when searching for a system tool. But, from
what I've read, apt is not a replacement for apt-get, and the
expectation is not that newer Debian systems will eventually have apt
instead of apt-get. The apt tool is intended as a more user friendly
interface, while apt-get remains a stabler UI and more complete toolset
better suited for scripting. For this particular script, we aren't doing
any parsing, so it's academic. But given that we're not relying on a
behavior difference between the two, I've switched the order to favor
apt-get to adhere to the spirit of their differences.

I'm actually not aware of an installation where you would have apt-get
without apt or vice-versa, so a fallback may be unnecessary as well, but
there is undoubtadely some edge case somewhere...