(gh-20) Download with curl --fail instead of --fail-with-body
The --fail-with-body parameter was introduced into curl in
in 7.76.0 - March 31 2021 (https://curl.se/changes.html).
Older vms we are still testing on may have curls predating this version.
For example, the rocky8/alma8 vm images have curl 7.61.1 - September 5
2018.

Switching the --fail provides basically the same semantics, but without
capturing the full response body in the file. Not trying to scrape the
respone, atm, though, so this should be fine.