(gh-20) Add a task to check the version of the puppet binary
Provides a consistent way of verifying the installed version for
acceptance testing.

Version comparison is handled by Ruby's Gem::Version class.
The agent's ruby is used to run the task.
If Bolt knows the target has the puppet-agent, then the check.rb is run.
If Bolt doesn't know whether puppet-agent is in place, the shell script
is run, which will check for the puppet binary and either fail or kick
off the check.rb if found.