(maint) Extract concrete inventory file for pr_testing workflow
The kvm_automation_tooling generated inventory files rely on a fork of the
puppetlabs-terraform module plugins to extract target information from
the Terraform state file
(jpartlow/puppetlabs-terraform@include-module-when-resolving-references).

This change came between v1 and v2 kvm_automation_tooling versions.

When we run openvox_bootstrap tasks, in CI, the Bolt installation only
has the base puppetlabs-terraform module to work with, so we need an
inventory file with explicit target information (ip addresses) now that
nested_vms is using kvm_automation_tooling@v2.

The .inventory.targets from 'bolt inventory show --format json' is
suitable as a concrete list of targets with a little massaging (there is
a 'groups' entry that Bolt doesn't expect in an inventory file).