(gh-26) Chown csr_attributes.yaml puppet:puppet on server
File.chown ignores nils, so if the puppet user/group do not exist,
nothing is done, and the file remains root:root which is correct for the
agent.

This chown is only important on the node running openvox-server , as
puppetserver may otherwise choke on first startup because the file can't
be read during its ca bootstrap. (I think because puppetserver runs as
the puppet user and is calling puppet ssl at some point.)

Note that it is the openvox-server package that creates the puppet
user/group at install.