(gh-35) Add a stop_service parameter to install task
This is for integration with openbolt for openvoxproject/openbolt@44.

Bolt was providing a stop_service parameter to the
puppetlabs-puppet_agent module here:
https://github.com/OpenVoxProject/openbolt/blob/b9bff5a8dfe2f41218281bf8c5e6ad901bcae460/lib/bolt/plugin.rb#L133

This was for puppetlabs/bolt#1204, and while the current linux packages,
at least, install with service stopped, it's probably still good
practice to ensure it for the openbolt case. (This may have been more of
an issue when some packages were still SysV?)

Since our install task is more general, it could be used for
openvox-server/openvoxdb, but there's no particular use case for this,
and I haven't bothered with acceptance tests for them.